requirements.txt
a2a-sdk>=0.3.0
click>=8.1.8
dotenv>=0.9.9
httpx>=0.28.1
langchain-google-genai>=2.1.4
langgraph>=0.4.1
pydantic>=2.11.4
python-dotenv>=1.1.0
sse-starlette>=2.3.5
starlette>=0.46.2
uvicorn>=0.34.2
ollama
pytest>=8.0.0
pytest-asyncio>=0.23.0
pytest-cov>=4.0.0